Fixed the build of froyo.
Since the addition of Gingerbread in korg remote, the master branch was for
gingerbread addition and have some API breakage (apache-http).

So, be sure to use the froyo branch for project located in korg remote by
removing the specified 'revision=master' for them.

stlport still use master branch (don't build with current NDK if switched)

Change-Id: I7af62b569aca26e545fd09f9b16486df10c29c05