Update patch set 1
Patch Set 1:

Arne -

Yes sir, performed "mka clobber" before build.  I also picked the following patch as well, http://review.cyanogenmod.com/#change,13906.  If you would like I can build without this second patch included and report back.

Also, the error above would appear when scrolling through some of the photos in the logs, but on screen couldn't tell there was any issue.  It wasn't until attempting to add the jpeg as a wallpaper that it was apparent there was an issue.  The screen to crop the wallpaper would just show an all grey image with a green line on very left edge.  This did not happen on every photo, was intermittent.

Patch-set: 1