Update patch set 1
Patch Set 1:

Arne -

I performed a clobber build with just this patch and experienced the same issue.  If possible, I would like someone else to build with this patch and watch the log while scrolling within various gallery albums to see if they experience something similar.  Unless they attempt to set one of the problem images(it's not all images) as wallpaper, they wouldn't know there was a problem with this patch.

Patch-set: 1