Update patch set 1
Patch Set 1:

To answer both questions: ICS support may happen, but it's unlikely. We expect the stragglers still on CM9 to move up to CM10 eventually, and no more stable CM9 releases are expected to happen.

As for builds that aren't CM, even if CM-based... no. The updater uses the get.cm infrastructure to distribute the builds, and to generate availability data. It is not written to accommodate random 3rd party notification URLs. There's a reasonable variety of other update software out there that can be used for non-CM builds, including ROM Manager (clockworkmod.com allows anyone to register and distribute builds, I've used it myself).

Patch-set: 1