Update patch set 1
Patch Set 1:

Greater configurability, in the source as well as in the config file's structure.

Plus an opportunity to increase/decrease the scope of thermald. Like mentioned, as newer devices are coming to market that have thermal daemons that control not just GPU/CPU but sensor polling rate, camera fps etc too based on thermal pressure. WE can add such features here too

The answer in short is "What is the advantage of CM over Stock ROM other than it is opensource ;) "

Patch-set: 1