Update patch set 4
Patch Set 4: Code-Review+1

With how many emails I have received asking for the binary, it feels like there is at least a small demand for including the openvpn binary in cm-11.0

GUI integration will take me quite a while since it is new territory for me. I'm not saying I won't give it a shot, but it will definitely take some time.

I vote for including the binary for now, at least to quell the influx of emails I am receiving asking for the 2.3.4 binary.

Patch-set: 4