Update patch set 1
Patch Set 1:

Even if you had the APK right now, you wouldn't be able to enable it unless you had a bunch of proprietary stuff (including headers needed to enable Listen support). I don't think we can ever add support for it to the OSS builds because of this.

Patch-set: 1