Update patch set 1
Patch Set 1:

Why not separate CAF HALs into 2 groups: A-Family and B-Family (or AF and BF as they have on their git)?  A-Family being 8960, 8064, and other legacy chipsets (8660, 7x30 I believe can be easily supported there) and B-Family being newer chips, such as 8974, 8084, 8226, 8610, and 8994.  It seems to make sense to keep similar chipsets together and is less work to maintain in the long run when new SoCs are added.

Patch-set: 1