Update patch set 1
Patch Set 1: Code-Review+1

Hello Ethan
It was just to make it consistent with the rest of the projects.
All the other projects using a different remote, place groups before remote and if there is a revision is it placed as the last attribute.

Patch-set: 1
Reviewer: Henning Nielsen Lund <3848@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=+1