Update patch set 1
Patch Set 1:

Not sure the difference between lz4 and lz4c.  In Debian 8.0 they both are from the same source package and they both seem to have the same command line arguments.

Note lz4 has two distinct file formats.  The "legacy" format is very simple but it's deprecated.  It is enabled with the "-l" (dash ell) option.  It used by the Linux kernel for ramdisks and it will be used in the upcoming transparent compression patches.

Patch-set: 1