Update patch set 1
Patch Set 1:

Exploit described at https://www.usenix.org/system/files/conference/woot15/woot15-paper-peles.pdf .  "This vulnerability allows for arbitrary code execution in the context of many apps and services and results in elevation of privileges. In this paper we also demonstrate a Proof-of-Concept exploit
against the Google Nexus 5 device, that achieves code execution inside the highly privileged system_server
process, and then either replaces an existing arbitrary application on the device with our own malware app
or changes the device’s SELinux policy. For some other devices, we are also able to gain kernel code execution
by loading an arbitrary kernel module. We had responsibly disclosed the vulnerability to Android Security Team which tagged it as CVE-2015-3825 (internally as ANDROID-21437603/21583894) and patched Android 4.4 / 5.x / M and Google Play Services."

The commit I am targeting to pull in is https://android.googlesource.com/platform/external/conscrypt/+/edf7055 , which seems to be the only difference between the r9 tag and r6.  This may not be preferred the way to do it, but I'm opening this for a place to get discussion rolling.  At the least, someone should probably put in a comment explaining the importance of that tag, and maybe a warning somewhere to clean it up when the default tag is updated to one that contains the commit.

Also it should be backported, but we can hash it out on CM-12.1 first.

Patch-set: 1