Update patch set 1
Patch Set 1:

I think it was presented on the 10th, or maybe the 11th.  The r9 tag is dated July 30th, so maybe Google published then, but the public was unaware of the severity.  The transparency and communication on this feels a bit lacking. :/  I guess Google were being subtle with the patch on purpose, but then that just hurt us when the embargo wasn't cleanly lifted.  Oh well, I know that's not yours or the CM community's fault, so thanks for your quick response!

Forking sounds like a lot of work for you, but having mixed tags is ugly too.  Replacing this patch sounds good to me.

Patch-set: 1