Update patch set 1
Patch Set 1:

stlport doesn't have a 6.0.0_r1 branch though. Also, would this be using the prebuilt stlport? If so it'd be better to fork and edit the whitelist as necessary, but I haven't touched stlport so I'm not sure what devices would even need to use it or if it truly is Nexus-specific.

Patch-set: 1