Update patch set 1
Patch Set 1:

Yeah, which is why no one is building the devices that use CAF HALs unless we already have the HALs for them. 8916 and 8994 are on my personal GitHub (https://github.com/Rashed97) if anyone wants to build those platforms. I'll upload the rest as soon as CAF pushes them.

Patch-set: 1