Update patch set 1
Patch Set 1:

toybox doesn't include a shell by default, and we need that for all Sony devices.
I tried enabling it, but it isn't working quite right, exits right after command finishes executing.

I would dig into that, but given that sh is a pending non default applet, I'll try to make a better program for the Sony devices than keep using the shell script we had.
If that isn't in time and this has to go, we could temporarily use a prebuilt busybox binaries (for bringup purposes)

Patch-set: 1