Update patch set 3
Patch Set 3:

Hi Dan, maybe you have noticed I typed I was considering to come back to CM after years. So, that might be a good reason for being here?

Besides that, I donot consider things delirious. I call it realistic to be critic on a review system where commits go in like "lets try and error".

So it is more about the proces that comes with this commit rather than what it implies.

But lets look to it from the bright side. For the first time in years I enjoyed CM today. Feels good!

Patch-set: 3