Update patch set 1
Patch Set 1:

Joey, It's been library we've used internally to stick some common custom widgets to share across projects. But with the addition of RuntimePermissions, it is suggested to throw up a SnackBar if the reason for the permission request is not clear. A number of OSS apps need this and, since the support/design SnackBar requires all AppCompat Activities/Styles, I've stripped out the compat dependencies and added to UICommon.

Patch-set: 1