Update patch set 1
Patch Set 1:

@Michael
That's what I used to initialize repo so I don't know why doing it again would change what is synced on github.  The only time I've seen this error was in France so I was assuming that it had something to do with the routing.

I can abandon it if you think I should though.

Patch-set: 1