Update patch set 15
Patch Set 15:

Anyone able to boot this stuff? I'm getting a ServiceManager crash: http://pastebin.com/UFURXwWz

And why aren't we starting with clean branches with CAF on top, like we've previously done? It would probably be better to leave all the CM specific stuff out for now, and merge that as we go, when it has been tested on CM14. Right now everything just seems like a complete mess.

Patch-set: 15