Update patch set 1
Patch Set 1:

Please let's not smoosh it all into a single repo (I think that's what is being suggested here). It creates a fucking *nightmare* of merge issues where Git tries to be smart and track file movement but it never gets it right because of duplicated files and such.

Patch-set: 1