Update patch set 1
Patch Set 1:

Thanks Adrian, you are right... 

This branch is already updated with the latest version of mr2. 

I will go ahead and merge the latest patches individually like I did for the stable branches if you are ok with that. 
We can plan to jump to MR3 when a new update gets released

Patch-set: 1