Moritz: Thanks for taking care about the security fixes for cm-13.0!
The android-6.0.1_r78 git tag just appeared after I pinged google via email.

Somehow I missed your earlier patches in gerrit, so I'm fine with abandoning my work in this area.

What's a bit odd: One of the fixes is in external/libopus (CVE-2017-0381). The output of "Found 13 updated projects in AOSP" didn't detect a change in that branch. Could that be because you've already have the backported patch in your branch?

IMHO it's better to merge the full repos, but that's something that needs to be done by somehow with direct git commit access. Pushing single patches for each repo via gerrit kills the history from AOSP.

No idea what the previous CM workflow for AOSP updates like this was.

Change-Id: Ia872c61c122b0c5ba75711c609023cfce17804f5