Update patch set 1
Patch Set 1:

> Why do we do this? Isn't the aosp one very limited with the things
 > you can do with it.

CMFM has been neglected for far too long. Lots of bugs and we'd rather focus on the OS than rewriting it at this point. Some changes have been made to the built-in manager so its a bit more functional, but if you need a full-fledged file manager, we'd recommend something third party.

Patch-set: 1