manifest: use updated svox tag to fix picotts
PicoTTS is currently broken in AOSP on 64bit nougat builds.
It was fixed in AOSP master and made it into oreo release, but never
made it into nougat release.
https://android-review.googlesource.com/302872
It is in an updated nougat preview tag though, so we can switch to it.

The svox repo can't be forked due to the following:
https://github.com/github/dmca/commit/1ec1226fd74dc5671807f0c89a25a3bdcfded5d7

BUGBASH-475

Change-Id: I2ab74ba6428614ef8824f52669c37b52daa59927