Update patch set 1
Patch Set 1: Code-Review-1

admin TODO list:
https://gist.github.com/sam3000/efdf9780e757f5436db435da1b36525f

Open question on the new name for CMParts ?   I propose LineageSettings (and have this in the gerrit).

As per slack, I have gerrit stacks to push up once we have projects and branches.  Most things in CMParts have to be disabled (I have a gerrit for that) but it allows features to be reviewed and brought up one at a time.

Patch-set: 1
Label: Code-Review=-1