Update patch set 1
Patch Set 1:

If we follow the rules we've used so far where we change cm -> lineage, cyanogen -> lineage and cyanogenmod -> lineageos we end up with paths like:
vendor/lineagesdk/packages/LineageSettingsProvider/src/org/lineageos/lineagesettings

And this is going to get confusing quickly since if we also have a package/apps/LineageSettings where files in this package are going to be importing lineageos.providers.LineageSettings for settings db lookups.

I'm beginning to think that LineageParts for the settings side might be a better choice to avoid confusion with lineagesdk settings provider.

Patch-set: 1