Update patch set 5
Patch Set 5:

The rebrand commits are now ready for review (pls ignore other topic:lineagesdk-oreo-bringup commits at present).

lineage-sdk:
https://review.lineageos.org/#/c/190470/ update paths
https://review.lineageos.org/#/c/190471/ update file contents

vendor/lineage:
https://review.lineageos.org/#/c/190472/ update paths
https://review.lineageos.org/#/c/190473/ update file contents

LineageParts:
https://review.lineageos.org/#/c/190474/ update paths
https://review.lineageos.org/#/c/190475/ update file contents


One point of note is that I began the rebrand using "cmsdk" -> "lineagesdk" because this was when I was using the path vendor/lineagesdk.  Of course the path changed to lineage-sdk but, for now, I've left the replacement string as lineagesdk because  this reads better to me and I wanted to confirm the direction before changing anything again.

Three options:
1) Leave the path as lineage-sdk but the file content strings as lineagesdk.  I'm sure no-one likes the inconsistency with this.
2) Change the "lineagesdk" strings to "lineage-sdk".
3) Change the project path of lineage sdk from lineage-sdk to lineagesdk (and update any path references).

Preferences ?

Patch-set: 5
Label: Verified=0