Update patch set 2
Patch Set 2:

We ran into this problem a few months ago too. Where the tag we're tracking is not updated to newest firmware for some devices.  It's because google only updates the firmware on the tag the device uses. Pain to deal with because it requires our manual intervention. Like in your example, r29 has newest 4354 firmware for fugu. And r31 has newest 4358 firmware for angler. ugh

Patch-set: 2