Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Shouldn't you just create a new branch instead?

That would be the other viable solution, but so far, every time we tracked a newer revision of a project, we did that through a static revision (using the revisions that were available on the server already).

iirc, zif once mentioned that branches are expensive (on memory, I guess?), but if it's the preferable solution despite of that, I guess we should use a branch.

Patch-set: 2
Reviewer: Gerrit User 15734 <15734@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>