lineage: Enable NFC NXP repos
Due to the fact that halimpl and hidimpl have diverged in Q,
(being that the sdm845 tag supports PN80T/PN81T/PN553/PN557 and
the sm8150 tag support only SN100x), it is now needed to have
different branches for the repositories. The devices must set the
soong namespace to select the proper HAL impl.

Change-Id: I431a1ca9445b3981f15e39236a97b6d7f7891b28