Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Did we merge any netd commits? Usually manifest entry goes in after we have merged commits in the repository

Nope, not yet.  But we know we'll need it soon and it's safe to switch now.  I'm not sure why we'd delay merging and contributing to "yet another open commit which could be merged".

What is the reason behind your preference to wait ?  (for a repo that we can be 100% sure we are forking)

Patch-set: 3